Implement hard-edged shadows for OpenCL

+++ src/GeometricObjects/Primitives/Sphere.cpp
@@ -109,18 +109,18 @@ Sphere::hit(const Ray& ray, double& tmin, ShadeRec& sr) const {
 
 bool
 Sphere::shadow_hit(const Ray& ray, float& tmin) const {
-	double 		t;
+	float 		t;
 	Vector3D	temp 	= ray.o - center;
-	double 		a 		= ray.d * ray.d;
-	double 		b 		= 2.0 * temp * ray.d;
-	double 		c 		= temp * temp - radius * radius;
-	double 		disc	= b * b - 4.0 * a * c;
+	float 		a 		= ray.d * ray.d;
+	float 		b 		= 2.0 * temp * ray.d;
+	float 		c 		= temp * temp - radius * radius;
+	float 		disc	= b * b - 4.0 * a * c;
 
 	if (disc < 0.0)
 		return(false);
 	else {
-		double e = sqrt(disc);
-		double denom = 2.0 * a;
+		float e = sqrt(disc);
+		float denom = 2.0 * a;
 		t = (-b - e) / denom;    // smaller root
 
 		if (t > kEpsilon) {
